Change default to fp16, set use_gpu to true for scripts

+++ examples/parser.py
@@ -81,7 +81,7 @@ class Parser(ArgumentParser):
 
     def add_execution_mode_config_arguments(self):
         self.add_argument("--backend", choices=["simulate", "pytorch"], required=True)
-        self.add_argument("--dtype", choices=["fp32", "fp16"], default="fp32")
+        self.add_argument("--dtype", choices=["fp32", "fp16"], default="fp16")
 
     def add_simulation_output_config_arguments(self):
         self.add_argument("--trace_file", type=str, default=None, help="Trace file")
